pyproject.toml
[project]
name = "hactl-companion"
version = "0.3.0"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-aiohttp>=1.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "openapi-spec-validator>=0.7",
]
integration = [
    "pytest>=8.0",
    "requests>=2.31",
    "websocket-client>=1.7",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["companion"]
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
